+++ SpellsPlugin/src/com/elmakers/mine/bukkit/plugins/spells/builtin/PhaseSpell.java
@@ -1,0 +1,50 @@
+package com.elmakers.mine.bukkit.plugins.spells.builtin;
+
+import org.bukkit.Material;
+
+import com.elmakers.mine.bukkit.plugins.nether.NetherManager;
+import com.elmakers.mine.bukkit.plugins.spells.Spell;
+
+public class PhaseSpell extends Spell
+{
+	public PhaseSpell(NetherManager nether)
+	{
+		this.nether = nether;
+	}
+	
+	@Override
+	public boolean onCast(String[] parameters)
+	{
+		if (nether == null)
+		{
+			return false;
+		}
+		return nether.go(player, parameters) != null;
+	}
+
+	@Override
+	protected String getName()
+	{
+		return "phase";
+	}
+
+	@Override
+	public String getCategory()
+	{
+		return "nether";
+	}
+
+	@Override
+	public String getDescription()
+	{
+		return "Phase between worlds";
+	}
+
+	@Override
+	public Material getMaterial()
+	{
+		return Material.GOLD_RECORD;
+	}
+
+	private NetherManager nether;
+}